reannotation potentially reflexive pronouns
Reannotation of potentially reflexive pronouns "me", "te", "nos", "os" and "se", as outlined in the article "Reflexive pronouns in Spanish Universal Dependencies", which was published in n°64 of the Procesamiento del Lenguaje Natural journal (ISSN 1135-5948.). Importantly, the reannotations proposed in the article were slightly modified following personal email correspondence with Dan Zeman, to obtain more consistent cross-lingual annotations. For example, the proposed changes to the verbal head of the pronoun were discarded.